fix slice bug, re organize code
